% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_iodine_deficiency.R
\name{detect_iodine}
\alias{detect_iodine}
\alias{detect_iodine_general}
\alias{detect_iodine_pregnant}
\alias{detect_iodine_lactating}
\title{Determining Iodine Status in Population}
\usage{
detect_iodine(x, group = c("general", "pregnant", "lactating"))

detect_iodine_general(x)

detect_iodine_pregnant(x)

detect_iodine_lactating(x)
}
\arguments{
\item{x}{Either the single median value of the urinary iodine (μg/L) of the
population of interest or the vector of urinary iodine (μg/L) values of the
population of interest.}

\item{group}{The population group of interest as described below:\tabular{ll}{
   \strong{Population} \tab \strong{Short-form} \cr
   School-age-children (6 years or odler) \tab "general" \cr
   Pregnant Women \tab "pregnant" \cr
   Lactating Mothers and Children under 2 years old \tab "lactating" \cr
}


This should be specified using the short-form of the population group of
interest. Default to \emph{"general"}.}
}
\value{
A character vector indicating the iodine intake status of the
population of interest.
}
\description{
Identification of population iodine intake using urinary iodine concentration
}
\examples{
 # Create testing data
 x <- runif(20, min = 10, max = 350)

 # For individual target group function;
  detect_iodine_general(x)   # school-age-children (6 years or above)
  detect_iodine_pregnant(x)  # pregnant
  detect_iodine_lactating(x) # lactating or U2 children


 # For overall population function;
  detect_iodine(x = x, group = "general")

  detect_iodine(x = x, group = "pregnant")

  detect_iodine(x = x, group = "lactating")

 # test with sample dataset
 dta <- subset(iodine, group == "Child")
 detect_iodine(x = dta$iodine, group = "general")

}
