% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_iron.R
\name{detect_iron_u5}
\alias{detect_iron_u5}
\title{Determining Iron Storage Status in individual and Population}
\usage{
detect_iron_u5(
  df,
  ferritin = NULL,
  sex = NULL,
  app = c("crp", "agp"),
  crp = NULL,
  agp = NULL
)
}
\arguments{
\item{x}{This parameter accepts either the single median value of the
interested population urinary iodine (μg/L) values or the vector which
contains urinary iodine (μg/L) values of the interest population.}

\item{group}{The cut-off point for population iodine intake based on the
urinary iodine value is different. Therefore, the parameter which require
to identify the type of interested population for analysis need to
mention in the function. The required population group for this parameter
is mentioned in the below table.\tabular{ll}{
   \strong{Population} \tab \strong{Value} \cr
   School-age-children (6 years or odler) \tab "general" \cr
   Pregnant Women \tab "pregnant" \cr
   Lactating Mothers and Children under 2 years old \tab "lactating" \cr
}


Please note that the group parameter only accepts a single value of the
specific group's name, not the combination of more than one groups' name
in vector form. (for example, group == "general")}
}
\value{
iodine   the iodine intake stauts of the study (sample) population
}
\description{
Identification of individual iron storage status on the basis of serum
ferritin concentration
}
\examples{
 # Create testing data
 x <- runif(20, min = 10, max = 350)

 # For individual target group function;
  detect_iodine_general(x) # school-age-children (6 years or above)
  detect_iodine_pregnant(x) # pregnant
  detect_iodine_lactating(x) # lactating or U2 children


 # For overall population function;
  detect_iodine(x = x,
               group = "general")

  detect_iodine(x = x,
               group = "pregnant")

  detect_iodine(x = x,
               group = "lactating")

 # test with sample dataset
 dta <- subset(iodine_sample, group == "Child")
 detect_iodine(x = dta$iodine,
               group = "general")


}
