% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ferritin}
\alias{ferritin}
\title{Serum ferritin (µg/l) level sample dataset}
\format{
A data frame with 10 columns and 19449 rows.\tabular{ll}{
   \strong{Variables} \tab \strong{Description} \cr
   \emph{psu} \tab Primary sampling unit identifier \cr
   \emph{state} \tab State identifier \cr
   \emph{locality} \tab Locality identifier \cr
   \emph{sex} \tab Sex (1 = Male; 2 = Female) \cr
   \emph{m.age} \tab Age (in years) of female/mother with ages > 5 years \cr
   \emph{ch.age} \tab Age (in months) of children with ages less than 5 years \cr
   \emph{group} \tab Group classification \cr
   \emph{crp} \tab acute phase response proteine: C-reactive protein (CRP) value (mg/L) \cr
   \emph{ferritin} \tab Serum ferritin (μg/L) values \cr
   \emph{agp} \tab acute phase response proteine: α1-acid-glycoprotein (AGP) value (g/L) \cr
   \emph{age_group} \tab indicate study population (Under 5 or over 5 years old) \cr
}
}
\source{
From a micronutrient survey in a country in East Africa
}
\usage{
ferritin
}
\description{
This sample dataset contains 19449 observations and 10 variables.
}
\examples{
# explore the first 20 observations from the dataset
head(ferritin, 20)

}
\keyword{datasets}
