% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ferritin_sample}
\alias{ferritin_sample}
\title{Urinary iodine (μg/L) values for women and children}
\format{
A data frame with 8 columns and 312 rows.\tabular{ll}{
   \strong{Variables} \tab \strong{Description} \cr
   \emph{psu} \tab Primary sampling unit identifier \cr
   \emph{state} \tab State identifier \cr
   \emph{locality} \tab Locality identifier \cr
   \emph{sex} \tab Sex (1 = Male; 2 = Female) \cr
   \emph{m.age} \tab Age (in years) of female/mother with ages > 5 years \cr
   \emph{ch.age} \tab Age (in months) of children with ages less than 5 years \cr
   \emph{group} \tab Group classification \cr
   \emph{crp} \tab acute phase response proteine: C-reactive protein (CRP) value (mg/L) \cr
   \emph{ferritin} \tab Serum ferritin (μg/L) values \cr
   \emph{agp} \tab acute phase response proteine: α1-acid-glycoprotein (AGP) value (g/L) \cr
}
}
\source{
will update later!
}
\usage{
ferritin_sample
}
\description{
This sample dataset (\code{iodine_sample}) contained 312 observations and 8 variables.
The detailed variable's description are mentioned in the following table.
}
\examples{
# explore the first 6 observations from the dataset
head(ferritin_sample)

}
\keyword{datasets}
