% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_undernutrition.R
\name{find_child_stunting}
\alias{find_child_stunting}
\title{Find children who are stunting/stunted from an anthropometric dataset}
\usage{
find_child_stunting(df, zscore = "haz", flag = NULL, add = TRUE)
}
\arguments{
\item{df}{A data.frame containing anthropometric z-score indices for
\verb{height-for-age}}

\item{zscore}{A character value for variable name for \verb{height-for-age z-score}
in \code{df}. Default is \code{haz}}

\item{flag}{A character value for variable name of coded flags applied to
\verb{height-for-age z-score} values in \code{df}. Default is NULL indicating
that flags have already been applied and flagged values have been censored.}

\item{add}{Logical. Should recode values for stunting/stunted be added to
\code{df}? Default is TRUE}
}
\value{
if \code{add} TRUE, returns \code{df} with recode values indicating stunting
or stunted cases.
}
\description{
Find children who are stunting/stunted from an anthropometric dataset
}
\examples{

find_child_stunting(df = zscorer::anthro1, zscore = "haz", flag = NULL, add = TRUE)

}
