% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flag_anaemia}
\alias{flag_anaemia}
\title{Anaemia Diagnosis}
\usage{
flag_anaemia(
  df,
  age = NULL,
  hb = NULL,
  sex = NULL,
  pregnant = NULL,
  add = TRUE
)
}
\arguments{
\item{df}{Survey dataset (as an R data.frame) with the following variables
present; sex, age, haemoglobin value, and pregnancy status. If the study
sample population did not contain the pregnancy population, the pregnancy
status (dummy variable yes/no) might not be required for this function.
But, the other variables are compulsory for function processing. If the
dataset did not contain those variables, please perform data processing
to have those variables in the dataset.}

\item{age}{Sample observation's age from the dataset. This variable should
be recorded in the year format. If your dataset has only an age variable
with month, please converts it into year format before using this function.}

\item{hb}{Sample observation's haemoglobin level from the dataset. The Hb
values should record in the grams per liter (g/L) units.}

\item{sex}{Sample observation's sex identifies from the dataset. This
variable should be coded in the dummy variable as \code{1} for "male" and \code{0}
for "female". If your dataset's respondent sex variable was not recorded
this way, please make necessary data recoding before using this function.}

\item{pregnant}{Sample observation's pregnancy status from the dataset. This
variable should be coded as a dummy variable: \code{1} for pregnant and \code{0}
for non-pregnant. Please perform the necessary data processing for
recoding if your dataset observation's pregnancy status was not in dummy
variable format.}

\item{add}{This parameter's default option is TRUE and will add new
generated variables \code{anaemia_all} to your existing dataset applied in
this function. This newly developed categorical variable comprises three
types of flag categories resulting from data checking results; no anaemia,
mild anaemia, moderate anaemia, and severe anaemia. The following table
explains the cut-off points applied in this diagnostic function.\tabular{llll}{
   \strong{Population} \tab \strong{Mild} \tab \strong{Moderate} \tab \strong{Severe} \cr
   Children 6-59 months of age \tab 100 - 109 \tab 70 - 99 \tab < 70 \cr
   Children 5-11 years of age \tab 110 - 114 \tab 80 - 109 \tab < 80 \cr
   Children 12-14 years of age \tab 110 - 119 \tab 80 - 109 \tab < 80 \cr
   Non-pregnant women \tab  \tab  \tab  \cr
   (15 years and above) \tab 110 - 119 \tab 80 - 109 \tab < 80 \cr
   Pregnant women \tab 100 - 109 \tab 70 - 99 \tab < 70 \cr
   Men \tab  \tab  \tab  \cr
   (15 years and above) \tab 110 - 129 \tab 80 - 109 \tab < 80 \cr
}}
}
\value{
A data frame with the same structure as \code{df}. The new variable
\code{anaemia_all} can be observed, containing the information the type of
anaemia present: mild, moderate, or severe.
}
\description{
Perform anaemia diagnosis based on age, sex and pregnancy status
}
