% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flag_anaemia_2}
\alias{flag_anaemia_2}
\title{Anaemia Diagnosis}
\usage{
flag_anaemia_2(df, pop_group = NULL, hb = NULL, add = TRUE)
}
\arguments{
\item{df}{Survey dataset (as an R data.frame) with the following variables
present; sex, age, haemoglobin value, and pregnancy status. If the study
sample population did not contain the pregnancy population, the pregnancy
status (dummy variable yes/no) might not be required for this function.
But, the other variables are compulsory for function processing. If the
dataset did not contain those variables, please perform data processing
to have those variables in the dataset.}

\item{pop_group}{This variable contains information about the characteristic
of observation required to identify the anaemia status. In anaemia diagnosis,
Hb thresholds are different among different populations, such as age group,
gender, and special condition (like pregnancy). Therefore, one variable
can indicate the type of sample population characteristic requirement for
anaemia diagnosis. The required population group for this variable is
mentioned in the below table.\tabular{ll}{
   \strong{Population} \tab \strong{Valye} \cr
   Children 6-59 months of age \tab 1 \cr
   Children 5-11 years of age \tab 2 \cr
   Children 12-14 years of age \tab 3 \cr
   Non-pregnant women (15 years and above) \tab 4 \cr
   Pregnant women \tab 5 \cr
   Men (15 years and above) \tab 6 \cr
}


if the dataset did not include this variable yet, please create one before
using this function. Sample r code for generating this variable is
mentioned below using the sample dataset \code{hbData}.

df$pop_group <- with(df,ifelse(
m.age >= 6 & m.age < 60, 1, ifelse(
ch.age >= 5 & ch.age < 12, 2, ifelse(
ch.age >= 12 & ch.age < 15, 3, ifelse(
ch.age >= 15 & !is.na(ch.age) & sex == 2, 4, ifelse(
group == "Pregnant Not Principal Carer" |
group == "Pregnant and lactating Principal Carer" |
group == "Pregnant Principal carer", 5, ifelse(
ch.age >= 15 & !is.na(ch.age) & sex == 1, 6, NA)))))))}

\item{hb}{Sample observation's haemoglobin level from the dataset. The Hb
values should record in the grams per liter (g/L) units.}

\item{add}{This parameter's default option is TRUE and will add new
generated variables \code{anaemia_all} to your existing dataset applied in
this function. This newly developed categorical variable comprises three
types of flag categories resulting from data checking results; no anaemia,
mild anaemia, moderate anaemia, and severe anaemia. The following table
explains the cut-off points applied in this diagnostic function.\tabular{llll}{
   \strong{Population} \tab \strong{Mild} \tab \strong{Moderate} \tab \strong{Severe} \cr
   Children 6-59 months of age \tab 100 - 109 \tab 70 - 99 \tab < 70 \cr
   Children 5-11 years of age \tab 110 - 114 \tab 80 - 109 \tab < 80 \cr
   Children 12-14 years of age \tab 110 - 119 \tab 80 - 109 \tab < 80 \cr
   Non-pregnant women \tab  \tab  \tab  \cr
   (15 years and above) \tab 110 - 119 \tab 80 - 109 \tab < 80 \cr
   Pregnant women \tab 100 - 109 \tab 70 - 99 \tab < 70 \cr
   Men \tab  \tab  \tab  \cr
   (15 years and above) \tab 110 - 129 \tab 80 - 109 \tab < 80 \cr
}}
}
\value{
A data frame with the same structure as \code{df} is named
\code{anaemia}. In this new data.frame, the new variable \code{anaemia_all} can be
observed, containing the information which observation was affected by
what type of anaemia: mild, moderate, or severe.
}
\description{
Perform anaemia diagnosis based on age, sex and pregnancy status
}
