% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_who.R
\name{flag_hlaz}
\alias{flag_hlaz}
\title{Apply World Health Organization (WHO) anthropometric z-score indices flagging
criteria to height- and/or length-for-age z-score}
\usage{
flag_hlaz(hlaz = NULL)
}
\arguments{
\item{hlaz}{A numeric value or vector of numeric values for
\emph{height-for-age z-score} (\emph{haz}) and/or \emph{length-for-age z-score} (\emph{laz}).}
}
\value{
A numeric value or vector of values of either \emph{0} or \emph{1} with a
value of \emph{0} indicating that z-score value is not flagged and a value of
\emph{1} indicating that z-score value is flagged.
}
\description{
Apply World Health Organization (WHO) anthropometric z-score indices flagging
criteria to height- and/or length-for-age z-score
}
\examples{
## Check if a single height-for-age or length-for-age z-score value is
## within WHO recommended limits
flag_hlaz(zscorer::anthro1$haz[1])

## Check if a vector of height-for-age and length-for-age z-score values are
## within WHO recommended limits
flag_hlaz(zscorer::anthro1$haz[1])

}
\author{
Ernest Guevarra
}
