% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_who.R
\name{flag_waz}
\alias{flag_waz}
\title{Apply World Health Organization (WHO) anthropometric z-score indices flagging
criteria to weight-for-age z-score}
\usage{
flag_waz(waz = NULL)
}
\arguments{
\item{waz}{A numeric value or vector of numeric values for
\emph{weight-for-age z-score} (\emph{waz}).}
}
\value{
A numeric value or vector of values of either \emph{0} or \emph{4} with a
value of \emph{0} indicating that z-score value is not flagged and a value of
\emph{4} indicating that z-score value is flagged.
}
\description{
Apply World Health Organization (WHO) anthropometric z-score indices flagging
criteria to weight-for-age z-score
}
\examples{
## Check if a single weight-for-age z-score value is within WHO recommended
## limits
flag_waz(zscorer::anthro1$waz[1])

## Check if a vector of weight-for-age z-score values are within WHO
## recommended limits
flag_waz(zscorer::anthro1$waz)

}
\author{
Ernest Guevarra
}
