% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_who.R
\name{flag_whlz}
\alias{flag_whlz}
\title{Apply World Health Organization (WHO) anthropometric z-score indices flagging
criteria to weight-for-height and/or weight-for-length z-score}
\usage{
flag_whlz(whlz = NULL)
}
\arguments{
\item{whlz}{A numeric value or vector of numeric values for
\emph{weight-for-height z-score} (\emph{whz}) and/or
\emph{weight-for-length z-score} (\emph{wlz}).}
}
\value{
A numeric value or vector of values of either \emph{0} or \emph{2} with a
value of \emph{0} indicating that z-score value is not flagged and a value of
\emph{2} indicating that z-score value is flagged.
}
\description{
Apply World Health Organization (WHO) anthropometric z-score indices flagging
criteria to weight-for-height and/or weight-for-length z-score
}
\examples{
## Check if a single weight-for-height or weight-for-length z-score value is
## within WHO recommended limits
flag_whlz(zscorer::anthro1$whz[1])

## Check if a vector of weight-for-height and weight-for-length z-score
##values are within WHO recommended limits
flag_whlz(zscorer::anthro1$whz)

}
\author{
Ernest Guevarra
}
