% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-correct_haemoglobin.R
\name{get_altitude_correction}
\alias{get_altitude_correction}
\title{Determine altitude correction factor for haemoglobin}
\usage{
get_altitude_correction(alt = NULL)
}
\arguments{
\item{alt}{Altitude, in metres, above sea level}
}
\value{
A numeric value or numeric vector of correction factor/s for
haemoglobin based on altitude
}
\description{
Determine altitude correction factor for haemoglobin
}
\examples{
get_altitude_correction(mnData$altitude)

}
\author{
Ernest Guevarra
}
