% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-correct_haemoglobin.R
\name{get_smoking_correction}
\alias{get_smoking_correction}
\title{Determine smoking status correction factor for haemoglobin}
\usage{
get_smoking_correction(smoke = NULL)
}
\arguments{
\item{smoke}{Simple integer code for smoking status classification:
0 = non-smoker; 1 = any smoking or half packet up to less than 1 packet a
day; 2 = 1 up to less than 2 packets a day; 3 = 2 or more packets a day.}
}
\value{
A numeric value or numeric vector of correction factor/s for
haemoglobin based on smoking status
}
\description{
Determine smoking status correction factor for haemoglobin
}
\examples{
get_smoking_correction(smoke = 1)

}
\author{
Ernest Guevarra
}
