% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_chart.R
\name{get_wfh_zchart}
\alias{get_wfh_zchart}
\title{Extract weight-for-length/height z-score data}
\usage{
get_wfh_zchart(
  baseurl = "http://www.who.int/childgrowth/standards/",
  gender = c("boys", "girls")
)
}
\arguments{
\item{baseurl}{Character value for the base URL of the WHO Child Growth
Standards. The default is http://www.who.int/childgrowth/standards/.
This can be adjusted should WHO change the base URL for the Child Growth
Standards section of their website.}

\item{gender}{A character value or vector that specifies the sex-specific
reference data to extract. The default is \code{c("boys", "girls")}.}
}
\value{
A data frame in tidy format with 4 coloumns and 21636 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{sex} \tab Sex of child. 1 = boy; 2 = girl \cr
   \emph{lh} \tab Length/height of child in centimetres \cr
   \emph{sd_type} \tab Type of z-score. Can be one of \emph{4SD}, \emph{3SD}, \emph{2SD}, \emph{1SD}, \emph{0}, \cr
     \tab \emph{-1SD}, \emph{-2SD}, \emph{-3SD}, \emph{-4SD} \cr
   \emph{sd_value} \tab Weight (kgs) value for the specified \code{sd_type} \cr
}
}
\description{
Function to extract weight-for-length/height z-score data from the World
Health Organization (WHO) website on Child Growth Standards
(http://www.who.int/childgrowth/standards/) which contains a
collection of reference data from the WHO Multicentre Growth Reference Study
(MGRS). This function extracts the expanded data provided by WHO for creating
growth charts for weight-for-length/height z-scores.
}
\examples{
# Get weight-for-length/height z-score expanded tables for charting
#if (interactive) get_wfh_zchart()


}
\author{
Ernest Guevarra
}
