% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hbData}
\alias{hbData}
\title{Serum haemoglobin values for women and children}
\format{
A data frame with 9 columns and 18487 rows.\tabular{ll}{
   \strong{Variables} \tab \strong{Description} \cr
   \emph{psu} \tab Primary sampling unit identifier \cr
   \emph{state} \tab State identifier \cr
   \emph{locality} \tab Locality identifier \cr
   \emph{sex} \tab Sex (1 = Male; 2 = Female) \cr
   \emph{m.age} \tab Age (in years) of female/mother with ages > 5 years \cr
   \emph{ch.age} \tab Age (in months) of children with ages less than 5 years \cr
   \emph{group} \tab Group classification \cr
   \emph{hb} \tab Haemoglobin values (g/L) \cr
   \emph{altitude} \tab Altitude of sampling location above sea level (metres) \cr
}
}
\source{
will update later!
}
\usage{
hbData
}
\description{
This sample dataset (\code{hbData}) contained 18,487 observations and 9 variables.
The detailed variable's description are mentioned in the following table.
}
\examples{
# explore the first 6 observations from the dataset
head(hbData)

}
\keyword{datasets}
