% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micronutr.R
\docType{package}
\name{micronutr}
\alias{micronutr-package}
\alias{micronutr}
\title{Utilities for Calculating Indicators of Vitamin and Mineral Status of
Populations}
\description{
Vitamin and mineral deficiencies continue to be a significant public health
problem. This is particularly critical in developing countries where
deficiencies to Vitamin A, iron, iodine, and other micronutrients lead to
adverse health consequences. Cross-sectional surveys are helpful in answering
questions related to the magnitude and distribution of deficiencies of
selected vitamins and minerals. This package provides tools for
calculating and determining select vitamin and mineral deficiencies using R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nutriverse.io/micronutr/}
  \item \url{https://github.com/nutriverse/micronutr}
  \item Report bugs at \url{https://github.com/nutriverse/micronutr/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Nicholus Tint Zaw \email{nicholustintzaw@gmail.com} [copyright holder]
}

}
\keyword{internal}
