% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnData}
\alias{mnData}
\title{Micronutrient survey data}
\format{
A tibble with 9 columns and 19449 rows.

**Variables** | **Description**
:--- | :---
*psu* | Primary sampling unit identifier
*sex* | Sex (1 = Male; 2 = Female)
*age* | Age (in years)
*group* | Group classification
*hb* | Haemoglobin values (g/L)
*calcium* | Serum calcium values ()
*crp* | Serum c-reactive protein (µg/l)
*ferritin* | Serum ferritin ()
*iodine* | Urine iodine ()
*altitude* | Altitude of sampling location above sea level (metres)
}
\source{
From a micronutrient survey in a country in East Africa
}
\usage{
mnData
}
\description{
This sample dataset is from a micronutrient survey in a country in East
Africa with 19449 observations and 9 variables described below.
}
\examples{
mnData

}
\keyword{datasets}
