% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutricheckr.R
\docType{package}
\name{nutricheckr}
\alias{nutricheckr}
\alias{nutricheckr-package}
\title{Tools for Nutrition Assessment and Analysis}
\description{
Nutrition assessment involves various methods and techniques. This
    package provides tools to support the implementation of these methods and
    techniques.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nutriverse/nutricheckr/}
  \item \url{https://github.com/nutriverse/nutricheckr}
  \item Report bugs at \url{https://github.com/nutriverse/nutricheckr/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

Authors:
\itemize{
  \item Nicholus Tint Zaw \email{nicholustintzaw@gmail.com}
}

}
\keyword{internal}
