% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_iodine.R
\name{detect_iodine}
\alias{detect_iodine}
\alias{detect_iodine_general}
\alias{detect_iodine_pregnant}
\alias{detect_iodine_lactating}
\title{Determining Iodine Status in Population}
\usage{
detect_iodine(x, group = c("general", "pregnant", "lactating"))

detect_iodine_general(x)

detect_iodine_pregnant(x)

detect_iodine_lactating(x)
}
\arguments{
\item{x}{a vector which contains urinary iodine (μg/L) values.}

\item{group}{The cut-off point for population iodine intake based on the
urinary iodine value is different. Therefore, the parameter which require
to identify the type of interested population for analysis need to
mention in the function. The required population group for this parameter
is mentioned in the below table.\tabular{ll}{
   \strong{Population} \tab \strong{Value} \cr
   School-age-children (6 years or odler) \tab "general" \cr
   Pregnant Women \tab "pregnant" \cr
   Lactating Mothers and Children under 2 years old \tab "lactating" \cr
}}
}
\value{
iodine   the iodine intake stauts of the study (sample) population
}
\description{
Identification of population iodine intake using urinary iodine concentration
}
\examples{
 # Create testing data
 x <- runif(20, min = 10, max = 350)

 # For individual target group function;
  detect_iodine_general(x) # school-age-children (6 years or above)
  detect_iodine_pregnant(x) # pregnant
  detect_iodine_lactating(x) # lactating or U2 children


 # For overall population function;
  detect_iodine(x = x,
               group = c("general", "pregnant", "lactating"))

}
