% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_iron_deficiency.R
\name{detect_iron_deficiency_u5}
\alias{detect_iron_deficiency_u5}
\alias{detect_iron_deficiency_5over}
\alias{detect_iron_deficiency}
\alias{detect_iron_deficiency_qualitative}
\title{Determine iron storage status}
\usage{
detect_iron_deficiency_u5(ferritin = NULL)

detect_iron_deficiency_5over(ferritin = NULL)

detect_iron_deficiency(ferritin = NULL, group = c("u5", "5over"))

detect_iron_deficiency_qualitative(ferritin = NULL, inflammation = NULL)
}
\arguments{
\item{ferritin}{A numeric value or numeric vector of serum ferritin level in
micrograms per liter (µg/l).}

\item{group}{A character value specifying the population target group to
determine iron status from. Can be either for under 5 year old ("u5") or
5 years and over ("5over"). Default to "u5".}

\item{inflammation}{Logical value or vector. Is subject in inflammation or
not?}
}
\value{
A character value or character vector of iron status. Can be either
"iron deficiency" or "no iron deficiency".
}
\description{
Given serum ferritin values, determine iron storage status.
}
\examples{
 # Iron storage status based on CRP only
 ferritin_corrected <- correct_ferritin(
   crp = ferritin$crp, ferritin = ferritin$ferritin
 )
 detect_iron_deficiency(ferritin_corrected)

 # Iron storage status based on AGP only
 ferritin_corrected <- correct_ferritin(
   agp = ferritin$agp, ferritin = ferritin$ferritin
 )
 detect_iron_deficiency(ferritin_corrected)

 # Iron storage status based on CRP and AGP
 ferritin_corrected <- correct_ferritin(
   crp = ferritin$crp, agp = ferritin$agp, ferritin = ferritin$ferritin
 )
 detect_iron_deficiency(ferritin_corrected)

 # Iron storage status - qualitative
 detect_iron_deficiency_qualitative(
   ferritin = 3, inflammation = TRUE
 )
 detect_iron_deficiency_qualitative(
   ferritin = c(2, 3, 5), inflammation = c(TRUE, FALSE, TRUE)
 )

}
