% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_who.R
\name{flag_who}
\alias{flag_who}
\title{Apply World Health Organization (WHO) anthropometric z-score indices flagging
criteria}
\usage{
flag_who(df, haz = NULL, waz = NULL, whz = NULL, add = TRUE)
}
\arguments{
\item{df}{A data.frame containing anthropometric z-score indices for
\verb{height-for-age}, \verb{weight-for-age} and/or \verb{weight-for-height}}

\item{haz}{A character value indicating the variable name in \code{df} for the
\verb{height-for-age z-score}}

\item{waz}{A character value indicating the variable name in \code{df} for the
\verb{weight-for-age z-score}}

\item{whz}{A character value indicating the variable name in \code{df} for the
\verb{weight-for-height z-score}}

\item{add}{Logical. Should flag values be added to \code{df}. Default is TRUE.}
}
\value{
If \code{add} FALSE, returns a vector of \code{flag} coded values indicating
problematic measurements. if \code{add} TRUE, returns \code{df} with additional
column named \code{flag} containing coded values indicating problematic
measurements
}
\description{
Apply World Health Organization (WHO) anthropometric z-score indices flagging
criteria
}
\examples{
flag_who(df = zscorer::anthro1, haz = "haz", waz = "waz", whz = "whz")

}
\author{
Ernest Guevarra
}
