% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnData}
\alias{mnData}
\title{Micronutrient survey data}
\format{
A tibble with 9 columns and 19449 rows.\tabular{ll}{
   \strong{Variables} \tab \strong{Description} \cr
   \emph{psu} \tab Primary sampling unit identifier \cr
   \emph{sex} \tab Sex (1 = Male; 2 = Female) \cr
   \emph{age} \tab Age (in years) \cr
   \emph{group} \tab Group classification \cr
   \emph{hb} \tab Haemoglobin values (g/L) \cr
   \emph{calcium} \tab Serum calcium values () \cr
   \emph{crp} \tab Serum c-reactive protein (microgram/l) \cr
   \emph{ferritin} \tab Serum ferritin () \cr
   \emph{iodine} \tab Urine iodine () \cr
   \emph{altitude} \tab Altitude of sampling location above sea level (metres) \cr
}
}
\source{
From a micronutrient survey in a country in East Africa
}
\usage{
mnData
}
\description{
This sample dataset is from a micronutrient survey in a country in East
Africa with 19449 observations and 9 variables described below.
}
\examples{
mnData

}
\keyword{datasets}
