% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutricheckr.R
\docType{package}
\name{nutricheckr}
\alias{nutricheckr}
\alias{nutricheckr-package}
\title{Tools for Nutrition Assessment and Analysis}
\description{
Nutrition assessment involves various methods and techniques. This
    package provides tools to support the implementation of these methods and
    techniques.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (0000-0002-4887-4415)

}
\keyword{internal}
